reorganise monitoring functions
